requirements.txt
sentence-transformers
groq
fastapi
uvicorn
pydantic
redis
qdrant-client
python-dotenv
